backend/requirements.txt
asgiref==3.8.1
distlib==0.3.9
Django>=4.2
django-cors-headers==4.7.0
django-vite==3.1.0
djangorestframework
djangorestframework_simplejwt==5.5.0
filelock==3.18.0
mysqlclient
platformdirs==4.3.8
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dotenv==1.1.0
pytz==2025.2
flask
flask_cors
sqlparse==0.5.3
virtualenv==20.31.2
sshtunnel
mysqlclient